--- app/src/main/java/com/norvellium/tasky/core/web/.gitignore
@@ -1,1 +1,0 @@
-/build
