+++ app/src/main/java/com/norvellium/tasky/core/web/TaskyApi.kt
@@ -1,10 +1,11 @@
 package com.norvellium.tasky.core.web
 
 import com.norvellium.tasky.BuildConfig
+import com.norvellium.tasky.core.web.response.NetworkResponse
+import com.norvellium.tasky.core.web.response.ResponseError
 import com.norvellium.tasky.data.remote.LoginBody
 import com.norvellium.tasky.data.remote.LoginResponse
 import com.norvellium.tasky.data.remote.RegistrationBody
-import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
