+++ app/src/main/java/com/norvellium/tasky/presentation/MainActivity.kt
@@ -5,7 +5,6 @@ import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import com.norvellium.tasky.R
-import com.norvellium.tasky.ui.main.MainViewModel
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
