+++ app/src/main/java/com/norvellium/tasky/presentation/MainViewModel.kt
@@ -1,4 +1,4 @@
-package com.norvellium.tasky.ui.main
+package com.norvellium.tasky.presentation
 
 import android.util.Log
 import androidx.lifecycle.SavedStateHandle
