+++ app/src/main/java/com/norvellium/tasky/presentation/login/LoginFragment.kt
@@ -1,7 +1,6 @@
-package com.norvellium.tasky.ui.login
+package com.norvellium.tasky.presentation.login
 
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
