+++ app/src/main/res/layout/fragment_register.xml
@@ -1,83 +1,90 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/Black"
-    tools:context=".ui.register.RegisterFragment">
+    tools:context=".presentation.register.RegisterFragment">
 
-    <LinearLayout
+    <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
+        android:layout_height="match_parent"
+        android:fillViewport="true">
 
-        <TextView
-            style="@style/Theme.Tasky.Styled.TextView.Inter.Regular"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_horizontal"
-            android:padding="50dp"
-            android:text="@string/create_your_account"
-            android:textColor="@color/White"
-            android:textSize="30sp" />
-
-        <RelativeLayout
+        <LinearLayout
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@drawable/rounded">
+            android:layout_height="wrap_content"
+            android:orientation="vertical">
 
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="vertical">
-
-                <EditText
-                    android:id="@+id/et_username"
-                    style="@style/Theme.Tasky.Styled.EditText.Inter.Regular"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:hint="@string/name" />
-
-                <EditText
-                    android:id="@+id/et_email_address"
-                    style="@style/Theme.Tasky.Styled.EditText.Inter.Regular"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:hint="@string/email_address" />
-
-                <EditText
-                    android:id="@+id/et_password"
-                    style="@style/Theme.Tasky.Styled.EditText.Inter.Regular"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:hint="@string/password" />
-            </LinearLayout>
-
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/btn_get_started"
-                style="@style/Theme.Tasky.Styled.Button.Inter.Regular"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_centerInParent="true"
-                android:text="@string/get_started" />
-
-            <FrameLayout
+            <TextView
+                style="@style/Theme.Tasky.Styled.TextView.Inter.Regular"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_alignParentStart="true"
-                android:layout_alignParentBottom="true">
+                android:layout_gravity="center_horizontal"
+                android:padding="50dp"
+                android:text="@string/create_your_account"
+                android:textColor="@color/White"
+                android:textSize="30sp" />
+
+            <RelativeLayout
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:background="@drawable/rounded">
+
+                <LinearLayout
+                    android:id="@+id/ll_registration_fields"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:orientation="vertical">
+
+                    <com.google.android.material.textfield.TextInputEditText
+                        android:id="@+id/et_username"
+                        style="@style/Theme.Tasky.Styled.EditText.Inter.Regular"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:hint="@string/name"
+                        android:inputType="textNoSuggestions" />
+
+                    <com.google.android.material.textfield.TextInputEditText
+                        android:id="@+id/et_email_address"
+                        style="@style/Theme.Tasky.Styled.EditText.Inter.Regular"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:hint="@string/email_address"
+                        android:inputType="textEmailAddress" />
+
+                    <com.google.android.material.textfield.TextInputEditText
+                        android:id="@+id/et_password"
+                        style="@style/Theme.Tasky.Styled.EditText.Inter.Regular"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:hint="@string/password"
+                        android:inputType="textPassword" />
+                </LinearLayout>
+
+                <androidx.appcompat.widget.AppCompatButton
+                    android:id="@+id/btn_get_started"
+                    style="@style/Theme.Tasky.Styled.Button.Inter.Regular"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:layout_below="@+id/ll_registration_fields"
+                    android:layout_marginTop="30dp"
+                    android:layout_marginBottom="75dp"
+                    android:text="@string/get_started" />
 
                 <ImageButton
                     android:id="@+id/btn_back"
                     android:layout_width="50dp"
                     android:layout_height="50dp"
-                    android:layout_gravity="center"
+                    android:layout_alignParentStart="true"
+                    android:layout_alignParentBottom="true"
                     android:background="@drawable/square_rounded_button"
-                    android:elevation="5dp"
-                    android:padding="5dp"
+                    android:padding="15dp"
+                    android:scaleType="centerCrop"
                     android:src="@drawable/ic_caret_back" />
 
-            </FrameLayout>
-        </RelativeLayout>
-    </LinearLayout>
+            </RelativeLayout>
+        </LinearLayout>
+    </ScrollView>
 </FrameLayout>
