+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/nav_login"
-        android:name="com.norvellium.tasky.ui.login.LoginFragment"
+        android:name="com.norvellium.tasky.presentation.login.LoginFragment"
         android:label="@string/login"
         tools:layout="@layout/fragment_login">
         <action
@@ -19,7 +19,7 @@
 
     <fragment
         android:id="@+id/nav_register"
-        android:name="com.norvellium.tasky.ui.register.RegisterFragment"
+        android:name="com.norvellium.tasky.presentation.register.RegisterFragment"
         android:label="@string/register"
         tools:layout="@layout/fragment_register">
     </fragment>
